+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Update `irx-config`
 - Fix register array derive regression
 - Sanitize `gen` keyword (new in Rust 2024 edition)
+- Add an `AVR` target which includes generic code to access configuration change protected registers in a more convenient way
 
 ## [v0.36.0] - 2025-03-09